Refatora códigos
