--- src/includes.hpp
@@ -1,9 +1,0 @@
-#include "defincludes.hpp"
-
-//
-//Usar #define/#undef NOMINMAX para conflitos com max() e min()
-//
-#define NOMINMAX
-#define WIN32_LEAN_AND_MEAN
-#include <Windows.h>
-#include <windowsx.h>
